clean up completed, added default css.
