--- assets/models/logo.gltf
@@ -1,91 +1,0 @@
-{
-	"asset":{
-		"generator":"Khronos glTF Blender I/O v4.0.43",
-		"version":"2.0"
-	},
-	"scene":0,
-	"scenes":[
-		{
-			"name":"Scene",
-			"nodes":[
-				0
-			]
-		}
-	],
-	"nodes":[
-		{
-			"mesh":0,
-			"name":"logo"
-		}
-	],
-	"meshes":[
-		{
-			"name":"logo",
-			"primitives":[
-				{
-					"attributes":{
-						"POSITION":0,
-						"NORMAL":1
-					},
-					"indices":2
-				}
-			]
-		}
-	],
-	"accessors":[
-		{
-			"bufferView":0,
-			"componentType":5126,
-			"count":6518,
-			"max":[
-				28.86751365661621,
-				30.618621826171875,
-				25
-			],
-			"min":[
-				-21.13202476501465,
-				-19.381378173828125,
-				-25
-			],
-			"type":"VEC3"
-		},
-		{
-			"bufferView":1,
-			"componentType":5126,
-			"count":6518,
-			"type":"VEC3"
-		},
-		{
-			"bufferView":2,
-			"componentType":5123,
-			"count":39096,
-			"type":"SCALAR"
-		}
-	],
-	"bufferViews":[
-		{
-			"buffer":0,
-			"byteLength":78216,
-			"byteOffset":0,
-			"target":34962
-		},
-		{
-			"buffer":0,
-			"byteLength":78216,
-			"byteOffset":78216,
-			"target":34962
-		},
-		{
-			"buffer":0,
-			"byteLength":78192,
-			"byteOffset":156432,
-			"target":34963
-		}
-	],
-	"buffers":[
-		{
-			"byteLength":234624,
-			"uri":"logo.bin"
-		}
-	]
-}
