--- assets/models/logo.mtl
@@ -1,2 +1,0 @@
-# Blender 4.0.0 MTL File: 'None'
-# www.blender.org
