+++ samples/style.css
@@ -75,39 +75,6 @@
     --main-shadow: 0 0 1rem -0.5rem hsla(30, 50%, 25%, 50%);
   }
   
-  body,
-  header,
-  main,
-  nav,
-  section,
-  article,
-  h1,
-  h2,
-  p,
-  a,
-  ul,
-  li,
-  time,
-  picture,
-  img,
-  video,
-  footer,
-  div,
-  span,
-  pre,
-  code {
-    display: block;
-    padding: 0;
-    margin: 0;
-    border: none;
-    /* outline: 1px solid #ff000033; */
-  }
-
-        /* html {
-          overflow: hidden;
-          overscroll-behavior: none;
-        } */
-  
   body {
     background-color: var(--c-plate);
     font-family: var(--font-body);
@@ -127,12 +94,6 @@
     margin: 0 auto;
   }
   
-  
-  #logo {
-    position: absolute;
-    inset: 0;
-  }
-  
   #mark {
     position: absolute;
     inset: 0;
@@ -149,21 +110,6 @@
     font-family: 'Roboto';
     line-height: 100%;
     font-size: 16px;
-  }
-
-  mr-app {
-    /* opacity:0.7; */
-  }
-
-  mr-app * {
-    box-sizing: border-box;
-    visibility:hidden;
-  }
-
-  mr-app > canvas {
-    position:fixed;
-    visibility:visible;
-    /* z-index: 1; */
   }
 
   .container {
@@ -187,7 +133,6 @@
     font-variation-settings: 'opsz' 80, 'wght' 550;
     font-weight: bold;
 
-    /* grid-row: 1; */
     grid-column: 2;
 
   }
