+++ src/index.js
@@ -1,3 +1,6 @@
+// CSS
+import './styles.css';
+
 // UTILS
 import './utils/extensions';
 
@@ -30,9 +33,7 @@ import './UI/Text/TextField';
 import './UI/Text/TextArea';
 
 // UI: LAYOUT
-import './entities/layout/Container';
-import './entities/layout/Row';
-import './entities/layout/Column';
+import './entities/Container';
 
 //   EXPORTS
 export * as THREE from 'three';
@@ -44,5 +45,3 @@ export { default as Entity } from './core/entity';
 
 // GEOMETRY
 export { default as UIPlane } from './geometry/UIPlane';
-// UI
-// export { default as Panel } from './UI/Panel';
